pyproject.toml
[tool.poetry]
name = "emergency_hack"
version = "0.1.0"
description = ""
authors = []
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "3.8.6"
sklearn = "^0.0"
pandas = "^1.2.4"
numpy = "*"
matplotlib = "^3.4.2"
scipy = "^1.6.3"
torch = "^1.8.1"
lightgbm = "^3.2.1"
catboost = "^0.25.1"
xgboost = "^1.4.2"
basemap = {url = "https://github.com/matplotlib/basemap/archive/master.zip"}
geopy = "^2.1.0"
tqdm = "^4.61.0"
transformers = "^4.6.1"
pytorch-lightning = "^1.3.3"
Cython = "^0.29.23"
hydra-core = "*"
hydra-optuna-sweeper = "*"
optuna = "^2.7.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
mypy = "^0.812"
flake8 = "^3.9.2"
isort = "^5.8.0"
autoflake = "^1.4"
pytest = "^6.2.4"
pre-commit = "^2.13.0"
black = "^21.5b1"
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | build
  | dist
  | notebooks
  | original
  | data
  )/
'''

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.8.1+cu111 torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html"
